@@ -23,18 +23,19 @@
         "docs": "https://tenancy.dev"
     },
     "require": {
-        "php": "^7.3|^8.0|^8.1",
+        "php": "^8.1|^8.2",
         "phpseclib/phpseclib": "~3.0",
         "doctrine/dbal": "~2.5|~3.0",
         "ramsey/uuid": "^4.0",
-        "laravel/framework": "^8.0|^9.0"
+        "laravel/framework": "^8.0|^9.0|^10.0"
     },
     "require-dev": {
         "fakerphp/faker": "^1.12",
         "laravel/laravel": "^8.0",
         "mockery/mockery": "^1.0",
         "phpunit/phpunit": "^9.0",
-        "symfony/dom-crawler": "~3.1"
+        "symfony/dom-crawler": "~3.1",
+        "rector/rector": "^0.15.16"
     },
     "autoload": {
         "psr-4": {
