@@ -38,7 +38,7 @@ class ConnectionSet extends AbstractEvent
      */
     public $purged;
 
-    public function __construct(Website $website = null, string $connection, bool $purged = true)
+    public function __construct(string $connection, Website $website = null, bool $purged = true)
     {
         $this->website = $website;
         $this->connection = $connection;
