@@ -22,36 +22,17 @@ use Illuminate\Contracts\Events\Dispatcher;
 class WebsiteUuidGeneration
 {
     /**
-     * @var Repository
-     */
-    private $config;
-    /**
-     * @var UuidGenerator
-     */
-    private $generator;
-
-    /**
      * WebsiteUuidGeneration constructor.
-     * @param Repository $config
-     * @param UuidGenerator $generator
      */
-    public function __construct(Repository $config, UuidGenerator $generator)
+    public function __construct(private readonly Repository $config, private readonly UuidGenerator $generator)
     {
-        $this->config = $config;
-        $this->generator = $generator;
     }
 
-    /**
-     * @param Dispatcher $events
-     */
     public function subscribe(Dispatcher $events)
     {
         $events->listen(Creating::class, [$this, 'addUuid']);
     }
 
-    /**
-     * @param Creating $event
-     */
     public function addUuid(Creating $event)
     {
         if (! $event->website->uuid && $this->config->get('tenancy.website.disable-random-id') !== true) {
